pyproject.toml
[tool.poetry]
name = "phone-bill"
version = "0.1.0"
description = "Prova backend CorpSystem"
authors = ["kaslufl leallucasfonseca@gmail.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
sqlalchemy = "^2.0.36"
fastapi-cli = "^0.0.5"
pydantic-settings = "^2.6.1"
alembic = "^1.14.0"
psycopg = {version = "^3.2.3", extras = ["binary"]}


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
pytest = "^8.3.3"
factory-boy = "^3.3.1"
testcontainers = "^4.8.2"
httpx = "^0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
